Add intoOfferLoading to cartLoading
