+++ client/state/data-layer/wpcom/sites/intro-offers/index.tsx
@@ -7,29 +7,6 @@ import {
 import { registerHandlers } from 'calypso/state/data-layer/handler-registry';
 import { http } from 'calypso/state/data-layer/wpcom-http/actions';
 import { dispatchRequest } from 'calypso/state/data-layer/wpcom-http/utils';
-import { IntroOffer } from 'calypso/state/sites/intro-offers/types';
-
-interface ResponseIntroOffer {
-	product_id: number;
-	product_slug: string;
-	currency_code: string;
-	formatted_price: string;
-	raw_price: number;
-}
-
-const mapResponse = ( {
-	product_id,
-	product_slug,
-	currency_code,
-	formatted_price,
-	raw_price,
-}: ResponseIntroOffer ): IntroOffer => ( {
-	productId: product_id,
-	productSlug: product_slug,
-	currencyCode: currency_code,
-	formattedPrice: formatted_price,
-	rawPrice: raw_price,
-} );
 
 const fetchIntroOffers = ( action: { siteId: number } ) => {
 	return http(
@@ -42,7 +19,7 @@ const fetchIntroOffers = ( action: { siteId: number } ) => {
 	);
 };
 
-const onUpdateSuccess = ( action: { siteId: number }, response: ResponseIntroOffer[] ) => {
+const onUpdateSuccess = ( action: { siteId: number }, response: unknown ) => {
 	return [
 		{
 			type: SITE_INTRO_OFFER_REQUEST_SUCCESS,
@@ -51,7 +28,7 @@ const onUpdateSuccess = ( action: { siteId: number }, response: ResponseIntroOff
 		{
 			type: SITE_INTRO_OFFER_RECEIVE,
 			siteId: action.siteId,
-			introOffers: response.map( mapResponse ),
+			payload: response,
 		},
 	];
 };
