+++ client/state/sites/intro-offers/types.ts
@@ -5,3 +5,17 @@ export interface IntroOffer {
 	formattedPrice: string;
 	rawPrice: number;
 }
+
+export interface ResponseIntroOffer {
+	product_id: number;
+	product_slug: string;
+	currency_code: string;
+	formatted_price: string;
+	raw_price: number;
+}
+
+export enum RequestStatus {
+	Pending = 'pending',
+	Success = 'success',
+	Failed = 'failed',
+}
